capi: address review - per-class file layout, Doxygen docs, trim unused overloads
Review fixes (PR #186, rjonaitis):

- File organization: public C headers move out of the c/ subdirectory
  into src/include/limesuiteng/, split per class (types_c.h,
  device_registry_c.h, sdrdevice_c.h, sdrdescriptor_c.h, rfstream_c.h,
  versioninfo_c.h) with an umbrella limesuiteng.h pairing the C++
  limesuiteng.hpp. Implementations split one-C-wrapper-per-C++-class
  under src/lib/, shared internals in a non-installed capi_private.h.
  The headers install via the existing public-header glob; the
  special-case install rule is removed.

- Documentation: all C API declarations rewritten as Doxygen
  (@brief/@param/@return, ///< members); file-top prose about the
  internal implementation and transition window removed from the
  public headers.

- SDRDevice: the 26 ChannelId delegating overloads with no callers are
  removed. ChannelId stays, as do SetAntenna(ChannelId, string) and
  GetAntennaName(ChannelId) - the antenna-by-name surface the C API is
  built on. The overloads were the C++-freeze migration target; under
  the C-ABI freeze direction they can return piecemeal when internal
  call sites actually migrate (non-virtual, purely additive).

Verified: build clean; 52/52 unit tests (2 removed with the trimmed
overloads); every public C header compiles standalone as C99 -Werror;
C99 smoke test links and passes; 32 lime_* exports unchanged.
