+++ src/include/limesuiteng/limesuiteng.h
@@ -1,0 +1,16 @@
+/**
+ * @file limesuiteng/limesuiteng.h
+ * @author Lime Microsystems
+ * @brief Umbrella header of the LimeSuiteNG C API.
+ */
+#ifndef LIMESUITENG_LIMESUITENG_C_H
+#define LIMESUITENG_LIMESUITENG_C_H
+
+#include "limesuiteng/types_c.h"
+#include "limesuiteng/device_registry_c.h"
+#include "limesuiteng/sdrdevice_c.h"
+#include "limesuiteng/sdrdescriptor_c.h"
+#include "limesuiteng/rfstream_c.h"
+#include "limesuiteng/versioninfo_c.h"
+
+#endif /* LIMESUITENG_LIMESUITENG_C_H */
